==> 02_scripts/preprocessing/05_cell_neighborhoods.rmd <==
---
title: "Map cellular neighborhoods across samples"
author: "Anthony R Cillo"
date: "March 6, 2025"
output: html_document
---

## Introduction

Identify spatially localized cell neighborhoods based on cell type annotations and spatial localization from the CosMx data.

## Load packages

```{r}

.libPaths("/ihome/acillo/arc85/Rlibs_Mar_2023")

suppressMessages({
    library(Seurat)
    library(tidyverse)
    library(patchwork)
    library(cluster)
    library(parallel)
    library(here)
})

```

## Load datasets

```{r}

ser_cos <- readRDS("../../03_output/SARC028_CosMx_object.rds")

ser_comb <- readRDS("../../03_output/SARC028_CosMx_and_scRNAseq_object_refined_annotations.rds")

```

## Filter cells and include cell annotations 

```{r}

## Excluded cells filtered in combined analysis
comb_meta <- ser_comb@meta.data %>%
    as_tibble(.,rownames="cb") %>%
    filter(data_type=="nanostring") %>%
    select(cb,final_cell_annot)

dim(ser_cos)

ser_cos <- ser_cos[,comb_meta$cb]

dim(ser_cos)

## Include cell annotations
meta <- ser_cos@meta.data %>%
    as_tibble(.,rownames="cb") %>%
    left_join(.,comb_meta,by="cb")

table(meta$final_cell_annot)

ser_cos@meta.data$final_cell_annot <- meta$final_cell_annot

## Combine B lineage and filter rare cell types
ser_cos@meta.data$final_cell_annot <- ifelse(ser_cos@meta.data$final_cell_annot %in% c("B_cells","Plasmablasts"),"B_cell_lineage",ser_cos@meta.data$final_cell_annot)

cell_types_include <- table(ser_cos@meta.data$final_cell_annot) %>%
    data.frame() %>%
    filter(Freq>9) %>%
    pull(Var1)

ser_cos <- ser_cos[,ser_cos@meta.data$final_cell_annot %in% cell_types_include]

table(ser_cos@meta.data$final_cell_annot)

```

## Function to find vector of neighbors

Define a function to identify the cells with a specific radius of other cells.

```{r}

cell_neighbor <- function(seurat_image,cell_types) {

    dat <- FetchData(seurat_image,vars="centroids")
    dat$cell_types <- cell_types

    dist_dat <- as.matrix(dist(dat[,1:2],method="euclidian"))

    neighbor_list <- vector("list",length=nrow(dat))

    for (i in 1:nrow(dat)) { 

        dat_sub <- dat
        dat_sub$cell_dist <- dist_dat[,i]

        dat_sub$include <- dat_sub$cell_dist<300

        neighbor_list[[i]] <- dat_sub[-i,] %>%
            filter(include==TRUE) %>%
            select(cell_types) %>%
            table() %>%
            prop.table() %>%
            as.vector()

    }

    neighbor_frame <- do.call(rbind,neighbor_list)
    colnames(neighbor_frame) <- levels(dat$cell_types)

    return(neighbor_frame)
    
}

```

## Cell neighborhoods across all images

```{r}

## Make sure cell types are a factor
ser_cos@meta.data$final_cell_annot <- as.factor(ser_cos@meta.data$final_cell_annot)

## Split into individual FOVs
ser_cos_split <- SplitObject(ser_cos,split="fov_selected")

## Loop over FOVs to ID cell neighorhoods
## Use parallel processing
set.seed(250306)
hoods <- mclapply(ser_cos_split,mc.cores=4,mc.set.seed=TRUE,function(x) {
    cell_neighbor(x$X1,cell_types=x@meta.data$final_cell_annot)
})

## Combine into dataframe
hood_frame_list <- vector("list",length=length(hoods))

for (i in 1:length(hoods)) {
    hood_frame_list[[i]] <- data.frame(hoods[[i]],fov=names(ser_cos_split)[[i]])
}

hoods_frame <- do.call(rbind,hood_frame_list)

```

## Evaluate spatial clustering solutions

```{r}

## Remove some cells that were not near other cells 
rows_exclude <- apply(hoods_frame,1,function(x) any(is.na(x)))

## Keep track of cells that get excluded
hoods_frame$cb <- colnames(ser_cos)

hoods_frame <- hoods_frame[!rows_exclude,]

## Evaluate clustering solutions
hoods_res <- vector("list",length=7)
hoods_dist <- dist(hoods_frame[,1:16])

for (i in 1:length(hoods_res)) {
    clust_res <- data.frame(hoods_frame,cluster=kmeans(hoods_frame[,1:16],centers=3+i)$cluster)
    hoods_res[[i]] <- silhouette(clust_res$cluster,dist=hoods_dist)
    print(paste("Finished ",i," of 7",sep=""))
}

clustering_results <- data.frame(cluster_number=seq(4,10,1),avg_silhouette_width=sapply(hoods_res,function(x) summary(x)$avg.width))

clustering_results %>%
    ggplot(.,aes(x=cluster_number,y=avg_silhouette_width)) +
    geom_col() +
    theme_bw()
## choosing 10 clusters

```

## Create spatial clustering with 10 clusters

```{r}

set.seed(250306)
hoods_clust <- data.frame(hoods_frame,cluster=kmeans(hoods_frame[,1:16],centers=10)$cluster)

mean_cell_freq <- hoods_clust %>%
    gather(cell_type,freq,-cluster,-fov,-cb) %>%
    group_by(cluster,cell_type) %>%
    summarize(mean_freq=mean(freq)) %>%
    ungroup() %>%
    spread(cell_type,mean_freq)

mean_cell_freq <- mean_cell_freq[,2:ncol(mean_cell_freq)] %>%
    data.frame()

pheatmap::pheatmap(mean_cell_freq,scale="column",color=colorRampPalette(colors = c("purple","black","yellow"))(1000))

## Renaming clustering to match published version
mapping <- data.frame(here_clust=seq(1,10,1),pub_clust=NA)

mapping[3,"pub_clust"] <- 8
mapping[9,"pub_clust"] <- 4
mapping[2,"pub_clust"] <- 5
mapping[5,"pub_clust"] <- 2
mapping[1,"pub_clust"] <- 1
mapping[10,"pub_clust"] <- 9
mapping[4,"pub_clust"] <- 7
mapping[7,"pub_clust"] <- 6
mapping[8,"pub_clust"] <- 3
mapping[6,"pub_clust"] <- 10

rownames(mean_cell_freq) <- paste("Hood ",c(mapping$pub_clust),sep="")

## Plotting
pheatmap::pheatmap(mean_cell_freq,scale="column",color=colorRampPalette(colors = c("purple","black","yellow"))(1000))

## Rename clusters in hood_clust
mapping

hood_clust <- hoods_clust %>%
    mutate(cluster=recode(cluster,
                          `1`=1,
                          `2`=5,
                          `3`=8,
                          `4`=7,
                          `5`=2,
                          `6`=10,
                          `7`=6,
                          `8`=3,
                          `9`=4,
                          `10`=9)
           )

```

## Add spatial clusters to Seurat object

```{r}

meta <- ser_cos@meta.data %>%
    as_tibble(.,rownames="cb")

meta <- left_join(meta,hoods_clust,by="cb")

ser_cos@meta.data$spatial_clusters <- meta$cluster

```

## Find DEGs across spatial neighborhoods

```{r}

ser_cos <- ser_cos[,!is.na(ser_cos@meta.data$spatial_clusters)]

Idents(ser_cos) <- "spatial_clusters"

hoods_degs <- FindAllMarkers(ser_cos,logfc.threshold=0.1)

degs_use <- hoods_degs %>%
    filter(p_val_adj<0.05) %>%
    filter(avg_log2FC>0) %>%
    select(gene) %>%
    distinct() %>%
    pull()

dat_use <- GetAssayData(ser_cos,slot="data",assay="SCT") %>%
    as.matrix()

ser_cos_genes <- cbind(ser_cos@meta.data,t(dat_use))

```

## Save data and session info

```{r}

# saveRDS(hoods_clust,file="../../03_output/SARC028_spatial_clustering_results.rds")
# saveRDS(hoods_degs,file="../../03_output/SARC028_CosMx_spatial_hood_DEGs.rds")
# saveRDS(ser_cos_genes,file="../../03_output/SARC028_CosMx_expression_metadata.rds")

sessionInfo()

```
